Move each monkeypatch to its own initializer file
I like this level of organization, rather than having everything thrown
together in a catch-all `config/initializers/monkeypatches.rb` file as
we were doing before.
